+++ app/App.js
@@ -1,23 +1,8 @@
 import React from 'react';
-import { StyleSheet, Text, View } from 'react-native';
-
-import Heading from './component/heading'
-import Body from './component/body'
+import DefaultScreen from './screen/common/default'
 
 export default function App() {
   return (
-    <View style={styles.container}>
-      <Heading />
-      <Body />
-    </View>
+    <DefaultScreen />
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
